[ADD] : 리뷰 평균값 업데이트 수정

+++ src/main/java/com/umc/goldenratio/api/domain/entity/Review.java
@@ -21,7 +21,7 @@ public class Review extends BaseTimeEntity {
     private String content;
 
     @Column(name = "review_score")
-    private BigDecimal score;
+    private double score;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "users_id")
